backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
rembg==2.0.56
Pillow==10.3.0
pdf2image==1.17.0
PyMuPDF==1.24.1
Werkzeug==3.0.2
python-docx==1.1.0
reportlab==4.2.0
openpyxl==3.1.5
python-pptx==1.0.2
yt-dlp==2023.12.30
